% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prediction_model.R
\name{build_prediction_model}
\alias{build_prediction_model}
\title{Build Prediction Model for Methylation Sites}
\usage{
build_prediction_model(
  BSobj,
  methInput,
  window_sizes,
  chunk1,
  chunk2,
  n_fold,
  cv_nesting,
  scaffoldIdentifier,
  outdir,
  record_runtime
)
}
\arguments{
\item{BSobj}{A BSseq object containing methylation data.}

\item{methInput}{A MethylationInput object containing SNP and methylation data.}

\item{window_sizes}{A vector of integers representing the sizes of windows surrounding methylation sites.}

\item{chunk1}{The starting index of methylation sites to process.}

\item{chunk2}{The ending index of methylation sites to process.}

\item{n_fold}{Integer, the number of folds for cross-validation.}

\item{cv_nesting}{Character, type of cross-validation nesting ("double" or "triple").}

\item{scaffoldIdentifier}{Character, a unique identifier for the scaffold being processed.}

\item{outdir}{Character, directory path to save the output files.}

\item{record_runtime}{Logical, indicating whether to record the runtime of the cross-validation.}
}
\value{
An object of class MethylationScaff containing the fitted models.
}
\description{
Constructs models to predict methylation as a function of single-nucleotide polymorphisms (SNPs)
across different window sizes surrounding methylation sites. The function processes a range of
methylation sites, fits a model for each, and compiles the results into a MethylationScaff object.
}
\examples{
\dontrun{
build_prediction_model(BSobj, methInput, c(1000, 2000), 1, 10, 5, "double", "scaffold1", tempdir(), TRUE)
}

}
