% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_cross_validation.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate Correlation Between Predicted and Observed Responses}
\usage{
calculate_correlation(fit, X_test, yhat_test)
}
\arguments{
\item{fit}{The fitted elastic net model object from 'glmnet'.}

\item{X_test}{Test set predictors, a subset of SNP matrix assigned for validation.
Dimensions: Observations (rows) x SNPs (columns).}

\item{yhat_test}{Numeric vector of observed response variable in the validation dataset.}
}
\value{
The correlation value between predicted and observed responses, or NA
if the variance of predictions is zero or NA.
}
\description{
Uses a fitted elastic net model to predict values and then computes the correlation
between these predicted values and the observed response values. This function is used
as a measure of model performance in methylome-wide association studies.
}
\examples{
\dontrun{
model_performance <- calculate_correlation(fitted_model, X_test, yhat_test)
}
}
