% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{clean_and_standardize_colnames}
\alias{clean_and_standardize_colnames}
\title{Clean and standardize column names}
\usage{
clean_and_standardize_colnames(summary_stats)
}
\arguments{
\item{summary_stats}{Data table of summary statistics}
}
\value{
Data table with standardized column names
}
\description{
Clean and standardize column names
}
