% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\name{convertToDataFrame}
\alias{convertToDataFrame}
\title{Convert MethylationModels to Data Frame}
\usage{
convertToDataFrame(object)
}
\arguments{
\item{object}{MethylationModels The \code{MethylationScaff} object to convert.}
}
\value{
data.frame Data frame representation of \code{MethylationScaff},
with column for each attribute except models and SNP weights.
}
\description{
Creates a data frame representation of \code{MethylationScaff}.
Each row corresponds to a \code{MethylationScaff}, with columns for each non-list attribute.
}
\examples{
\dontrun{
df <- convertToDataFrame(MethylationScaff)
}
}
