% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\name{initialize,MethylationGenome-method}
\alias{initialize,MethylationGenome-method}
\title{Constructor for MethylationGenome Class}
\usage{
\S4method{initialize}{MethylationGenome}(.Object, scaffoldIdentifiers, MethylationScaffs)
}
\arguments{
\item{scaffoldIdentifiers}{list List of scaffold identifiers.}

\item{MethylationScaffs}{list List of \code{MethylationScaff} objects.}
}
\value{
MethylationGenome object.
}
\description{
Initializes a \code{MethylationGenome} object with a list of scaffold identifiers and
corresponding \code{MethylationScaff} objects.
}
\examples{
\dontrun{
genome <- new("MethylationGenome", scaffoldIdentifiers = list("scaff1", "scaff2"),
MethylationScaffs = list(scaff1, scaff2))
}
}
