% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\name{initialize,MethylationInput-method}
\alias{initialize,MethylationInput-method}
\title{Constructor for MethylationInput Class}
\usage{
\S4method{initialize}{MethylationInput}(
  .Object,
  BSseq_obj,
  snp_data_path,
  no_cores = detectCores(),
  cov_path = NULL,
  start_site = NULL,
  end_site = NULL
)
}
\arguments{
\item{BSseq_obj}{A BSseq object containing methylation data.}

\item{snp_data_path}{A character string specifying the path to the directory containing SNP data in PLINK2 binary format.}

\item{no_cores}{Integer, the number of cores to use for parallel operations (defaults to available core count).}

\item{start_site}{Optional, the start site for subsetting methylation data based on indices in the file.}

\item{end_site}{Optional, the end site for subsetting methylation data based on indices in the file.}
}
\value{
An object of class MethylationInput, ready for methylation analysis.
}
\description{
Creates a new MethylationInput object by loading and processing methylation and SNP data from specified paths.
This method integrates methylation data from a BSseq object and SNP data in PLINK2 binary format, aligning
both data types alongside specified covariates for comprehensive analysis.
}
\examples{
# Example of initializing a MethylationInput object with methylation and SNP data
pgen_path = system.file("extdata", "chr1_sample_subset.pgen", package = "YourPackage")
data(chr1_methylation_sample_subset, package = "YourPackage")
methInput <- new("MethylationInput", BSseq_obj = BSseq_sample, snp_data_path = pgen_path)

}
