% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\name{initialize,MethylationScaff-method}
\alias{initialize,MethylationScaff-method}
\title{Constructor for MethylationScaff Class}
\usage{
\S4method{initialize}{MethylationScaff}(.Object, scaffoldIdentifier, models)
}
\arguments{
\item{scaffoldIdentifier}{character The identifier for the scaffold.}

\item{models}{list List of MethylationBase objects.}
}
\value{
MethylationScaff object.
}
\description{
Initializes a \code{MethylationScaff} object with a given identifier and a list of
\code{MethylationBase} models.
This is intended to store many \code{MethylationBase} models for the same scaffold,
typically a whole chromosome.
}
\examples{
\dontrun{
modelsList <- list(new("MethylationBase", ...), new("MethylationBase", ...))
scaff <- new("MethylationScaff", scaffoldIdentifier = "scaff1", models = modelsList)
}

}
