% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\name{loadSNPData}
\alias{loadSNPData}
\title{Helper function to load SNP data in PLINK 2 formats}
\usage{
loadSNPData(pvar_path, pgen_path, psam_path)
}
\arguments{
\item{pvar_path}{Path to the .pvar file}

\item{pgen_path}{Path to the .pgen file}

\item{psam_path}{Path to the .psam file}
}
\value{
A list containing pointers and data.tables for SNP data
}
\description{
Helper function to load SNP data in PLINK 2 formats
}
\examples{
\dontrun{
  snp_data <- loadSNPData("data/example.pvar", "data/example.pgen", "data/example.psam")
}
}
