% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\docType{class}
\name{MethylationGenome-class}
\alias{MethylationGenome-class}
\title{MethylationGenome Class}
\description{
Represents the genome-level structure in methylation analysis.
Contains multiple \code{MethylationScaff} objects, each associated with different scaffolds.
}
\section{Slots}{

\describe{
\item{\code{scaffoldIdentifier}}{list List of scaffold identifiers.}

\item{\code{MethylationScaffs}}{list List of MethylationScaff objects.}
}}

