% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\docType{class}
\name{MethylationScaff-class}
\alias{MethylationScaff-class}
\title{MethylationScaff Class}
\description{
Represents a scaffold structure in methylation analysis, containing multiple
\code{MethylationBase} models.
Each \code{MethylationScaff} is identified by a unique scaffold identifier.
}
\section{Slots}{

\describe{
\item{\code{scaffoldIdentifier}}{character Unique identifier for the scaffold.}

\item{\code{models}}{list List of MethylationBase objects.}
}}

