% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{mwas}
\alias{mwas}
\title{MWAS function}
\usage{
mwas(z, w, G)
}
\arguments{
\item{z}{Z-scores for effect of SNPs on external phenotype.}

\item{w}{Weights for effect of SNPs on methylation.}

\item{G}{SNP genotype matrix.}
}
\value{
Named vector containing z-score, p-value, and the number of weights.
}
\description{
Performs methylation-wide association study analysis.
}
