% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{MWASmodel}
\alias{MWASmodel}
\title{MWASmodel constructor}
\usage{
MWASmodel(methylationBase, summary_stats, mwas_out)
}
\arguments{
\item{methylationBase}{MethylationBase object}

\item{summary_stats}{Data table of summary statistics}

\item{mwas_out}{Numeric vector of MWAS output}
}
\value{
MWASmodel object
}
\description{
MWASmodel constructor
}
