% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{MWASresults}
\alias{MWASresults}
\title{MWASresults constructor}
\usage{
MWASresults(
  MWASmodels,
  pvar_path,
  pgen_path,
  psam_path,
  summary_stats_path,
  rds_path
)
}
\arguments{
\item{MWASmodels}{List of MWASmodel objects}

\item{pvar_path}{Path to pvar file}

\item{pgen_path}{Path to pgen file}

\item{psam_path}{Path to psam file}

\item{summary_stats_path}{Path to summary statistics file}

\item{rds_path}{Path to RDS file}
}
\value{
MWASresults object
}
\description{
MWASresults constructor
}
