% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{process_model}
\alias{process_model}
\title{Process a single MWAS model}
\usage{
process_model(methylationBase, my_SNPs, summary_stats)
}
\arguments{
\item{methylationBase}{MethylationBase object}

\item{my_SNPs}{SNP data}

\item{summary_stats}{Data table of summary statistics}
}
\value{
MWASmodel object
}
\description{
Process a single MWAS model
}
