% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{process_model_csv}
\alias{process_model_csv}
\title{Process a single MWAS model given weigths in matrix format}
\usage{
process_model_csv(this_cg_data, my_SNPs, summary_stats)
}
\arguments{
\item{this_cg_data}{matrix with six columns for chr, cg, snp, snp effect, A1, A2}

\item{my_SNPs}{SNP data}

\item{summary_stats}{Data table of summary statistics}
}
\value{
MWASmodel object
}
\description{
Process a single MWAS model given weigths in matrix format
}
