% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{process_MWAS_models}
\alias{process_MWAS_models}
\title{Process MWAS models}
\usage{
process_MWAS_models(
  my_rds,
  my_SNPs,
  paths,
  summary_stats_path,
  rds_path,
  summary_stats = NULL
)
}
\arguments{
\item{my_rds}{An object containing models}

\item{my_SNPs}{SNP data}

\item{paths}{List of paths to data files}

\item{summary_stats_path}{Path to summary statistics file}

\item{summary_stats}{Data table of summary statistics}
}
\value{
MWASresults object
}
\description{
Process MWAS models
}
