% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\name{regress_out_cov_parallel}
\alias{regress_out_cov_parallel}
\title{Regress trait (methylation) values on covariates and output residuals}
\usage{
regress_out_cov_parallel(
  methylations,
  cov_matrix,
  no_cores = detectCores() - 1
)
}
\arguments{
\item{methylations}{matrix of methylation rates, n x m for n samples, m sites}

\item{cov_matrix}{matrix of covariates, n x p for n samples, p covariates}
}
\value{
matrix of residuals, n x m for n samples, m sites
}
\description{
Regress trait (methylation) values on covariates and output residuals
}
\examples{
\dontrun{
adjusted_methylations <- regress_out_cov_parallel(methylations, cov_matrix)
}
}
