% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationModelClasses.R
\name{saveMethylationScaff}
\alias{saveMethylationScaff}
\alias{saveMethylationScaff,MethylationScaff-method}
\title{Save MethylationScaff Object}
\usage{
saveMethylationScaff(object, outputDir = "output/")

\S4method{saveMethylationScaff}{MethylationScaff}(object, outputDir = "output/")
}
\arguments{
\item{object}{MethylationScaff The MethylationScaff object to save.}

\item{outputDir}{character The directory to save the file, default is "output/".}
}
\description{
Saves a \code{MethylationScaff} object to an RDS file.
The file is named after the scaffold identifier and saved in the specified output directory.
}
