% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\name{validatePositionOverlap,MethylationInput-method}
\alias{validatePositionOverlap,MethylationInput-method}
\title{Validate Position Overlap Between Methylation Sites and SNPs}
\usage{
\S4method{validatePositionOverlap}{MethylationInput}(object, window_size)
}
\arguments{
\item{object}{MethylationInput object containing methylation positions and SNP data.}

\item{window_size}{Numeric value defining the window size around methylation sites to check for SNP overlap.}
}
\value{
Invisible. The function stops with an error message if there is no overlap within the specified window size.
}
\description{
This method checks if there is an overlap between methylation sites and SNP positions
given a specified window size. It ensures that the methylation data being used for analysis
has relevant SNP data within the defined proximity.
}
\details{
The function calculates the maximum and minimum positions for methylation sites and SNP data,
adjusted by the window size, to determine if any SNP falls within the vicinity of the methylation sites.
If not, then there's no overlap and the function stops with an error message.
}
\note{
This method is specific to MethylationInput objects. Ensure that your MethylationInput object
is properly initialized with valid methylation positions and SNP data before calling this method.
}
\examples{
# Assuming `methInput` is a MethylationInput object
validatePositionOverlap(methInput, 1000)

}
