% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\name{initialize,MethylationInput-method}
\alias{initialize,MethylationInput-method}
\title{Constructor for MethylationInput Class}
\usage{
\S4method{initialize}{MethylationInput}(.Object, BSseq_obj, snp_data_path, args)
}
\arguments{
\item{BSseq_obj}{BSseq The BSseq object containing methylation data.}

\item{snp_data_path}{character Path to the SNP data.}

\item{args}{list Additional arguments.}
}
\value{
MethylationInput object.
}
\description{
Initializes a MethylationInput object by processing methylation and SNP data.
}
\examples{
\dontrun{
methInput <- new("MethylationInput", BSseq_obj = BSseq_obj,
                 snp_data_path = "path/to/snp_data", args = list(...))
}
}
