% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylationInputClass.R
\docType{class}
\name{MethylationInput-class}
\alias{MethylationInput-class}
\title{MethylationInput Class}
\description{
This class encapsulates the input data processing for methylation analysis.
It includes methods for loading and processing methylation and SNP data, along with covariates.
}
\section{Slots}{

\describe{
\item{\code{methylations}}{Matrix of methylation data.}

\item{\code{genotype_IDs}}{Vector of genotype IDs.}

\item{\code{pvar1}}{Object representing variant information.}

\item{\code{pvar2}}{Data table of variant information.}

\item{\code{pgen}}{Object representing genotype information.}

\item{\code{psam}}{Data table of sample information.}

\item{\code{cov}}{Data frame of covariates.}
}}

