% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_mwas.R
\name{process_MWAS_models}
\alias{process_MWAS_models}
\title{Process MWAS models}
\usage{
process_MWAS_models(
  my_rds,
  my_SNPs,
  summary_stats,
  paths,
  summary_stats_path,
  output_path
)
}
\arguments{
\item{my_rds}{An object containing models}

\item{my_SNPs}{SNP data}

\item{summary_stats}{Data table of summary statistics}

\item{paths}{List of paths to data files}

\item{summary_stats_path}{Path to summary statistics file}

\item{output_path}{Path to output file}
}
\value{
MWASresults object
}
\description{
Process MWAS models
}
